backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pymysql
python-dotenv  # 可选，若你用它读取 env 文件
PyJWT